pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hems-nilm-gateway"
version = "0.1.0"
description = "NILM Gateway (M-GRU) für 1-Hz Smart-Meter Daten mit MQTT/Home-Assistant Integration."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Proprietary" }
authors = [
  { name = "Falk Haider" }
]
keywords = [
  "NILM",
  "HEMS",
  "Edge",
  "Gateway",
  "MQTT",
  "Home Assistant",
  "GRU",
  "PyTorch"
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "License :: Other/Proprietary License",
  "Operating System :: OS Independent"
]

dependencies = [
  "numpy>=1.23",
  "paho-mqtt>=1.6",
  "psutil>=5.9",
  "PyYAML>=6.0",
  "requests>=2.31",
  "torch>=2.0",
  "psycopg2-binary>=2.9"
]

[project.scripts]
hems-nilm-gateway = "hems_nilm_gateway.gateway.app:main"

[project.optional-dependencies]
dev = [
  "ruff>=0.4",
  "black>=24.0",
  "mypy>=1.7",
  "pytest>=7.4",
  "types-PyYAML>=6.0.12.12",
  "types-requests>=2.31.0.20240106"
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
include = ["hems_nilm_gateway*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
no_implicit_optional = true
check_untyped_defs = true
